requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
distlib==0.3.8
Django==4.2.11
django-cloudinary-storage==0.3.0
django-tinymce==4.1.0
filelock==3.16.1
idna==3.10
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==1.26.20
virtualenv==20.26.5
whitenoise==6.7.0
